Add files via upload
This is a variation of the files uploaded in the main task, but achieved with Object Oriented Programming Paradigm, rather than procedural programming. It involved adding some more features to the code from the main branch, and fully utilizing OOP method.
While this does fundamentally the same task as main branch, the codes differ in very significant ways but are also identical in some.